requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
SQLAlchemy==2.0.19
Flask-SQLAlchemy==3.0.5
openpyxl==3.1.2
reportlab
assertpy==1.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
# flask-jwt-extended==4.7.1
PyJWT==2.10.1
passlib==1.7.4
